+++ database/migrations/2025_10_28_120537_create_graves_table.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::create('graves', function (Blueprint $table) {
             $table->id();
-            $table->integer("cemetery_id");
-            $table->integer("latitude");
-            $table->integer("longitude");
+            $table->unsignedBigInteger("cemetery_id");
+            $table->unsignedInteger("latitude");
+            $table->unsignedInteger("longitude");
             $table->string("image_hash_url");
-            $table->integer("grave_number");
-            $table->integer("status_id");
+            $table->unsignedInteger("grave_number");
+            $table->unsignedBigInteger("status_id");
             $table->text("description")->nullable();
             $table->date("start_date");
             $table->date("end_date");
