+++ database/migrations/2025_11_14_122722_create_grave_users_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('grave_users', function (Blueprint $table) {
             $table->id();
-            $table->integer('grave_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('grave_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
         });
     }
